pyproject.toml
[project]
name = "pdf2text"
version = "0.1.0"
description = "A PDF to TXT extractor"
authors = [
    {name = "Martín Onetto"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn>=0.34.0,<0.35.0",
    "python-multipart>=0.0.5,<0.0.21",
    "pdf2image>=1.17.0,<2.0.0",
    "pdfminer-six>=20200517", 
    "easyocr>=1.7.2,<2.0.0",
    "numpy>=1.26.4,<2.0.0",     
    "scikit-learn>=1.3.0,<2.0.0",
    "pytesseract>=0.3.13,<0.4.0",
    "pymupdf>=1.26.0,<2.0.0"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.3"
ipykernel = "^6.29.5"
